feat:[SRM-13337]: Added nit fixes for Internal Change Source (#41594)
* feat:[SRM-13337]: Added nit fixes for Internal Change Source

* feat:[SRM-13337]: fixed PR tests

* feat:[SRM-13337]: Changed percentage value to return double

* feat:[SRM-13337]: Added fix for checkstyle

+++ srm-service/modules/cv-nextgen-service/service/src/main/java/io/harness/cvng/core/beans/change/ChangeSummaryDTO.java
@@ -24,6 +24,6 @@ public class ChangeSummaryDTO {
   public static class CategoryCountDetails {
     long count;
     long countInPrecedingWindow;
-    long percentageChange;
+    double percentageChange;
   }
 }
